Migrate to ruff 0.16
0.16 enables 413 rules by default instead of 59, so the bump has to take a
stance on the new ones. The scripts are one-shot build/harvest tools where a
few of the new rules fight the deliberate style, so those are ignored with the
reason spelled out; the rest of the findings are fixed.

Formatting is unchanged, so no reformat churn.

Fix #164

+++ scripts/build_sites_se.py
@@ -64,7 +64,7 @@ def locality_site(bbox, name):
         lat, lon = rec.get("decimalLatitude"), rec.get("decimalLongitude")
         if not lid or lat is None or lon is None:
             continue
-        radius = int(round(rec.get("coordinateUncertaintyInMeters") or 0)) or 100
+        radius = round(rec.get("coordinateUncertaintyInMeters") or 0) or 100
         return (
             lid,
             lat,
